refactor(mobile): extract embeddings queries out of db/queries.ts
db/queries.ts had grown to ~1660 lines. Move the on-device semantic-search
index queries (note_embeddings + task_embeddings, plus the display/lexical
helpers listWorkspaceIds/embeddingIndexStats/noteTitleById/workspaceIdForNote/
noteTextByIds) into db/embeddings-queries.ts (~320 lines), and factor the
shared LIVE / NOT_CONFLICT SQL fragments into db/sql.ts.

queries.ts re-exports the embeddings module (export * from "./embeddings-queries")
so every existing @/db/queries import keeps working unchanged — purely internal
reorganisation, no behaviour change. queries.ts drops to ~1350 lines.

Verified: mobile tsc + eslint . clean; mobile vitest project (26) green.
